Create MOHAMED EHAB
This code implements a FileManager class in C++ that provides static methods for persisting and loading Property objects to and from a text file. It integrates with a doubly circular linked list (DCLL<Property>) and assumes the existence of a Property class with methods like get_Property_ID(), get_neighborhood_Name(), and get_Price(), as well as a PropertyNode structure for the linked list.